requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.8
django-cors-headers==2.5.3
django-richtextfield==1.4.0
django-taggit==1.1.0
djangorestframework==3.9.2
elasticsearch==5.5.3
elasticsearch-dsl==5.1.0
idna==2.8
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
